renames only files

+++ rename_files.py
@@ -6,8 +6,9 @@ def rename_files(path: str, old: str, new: str) -> None:
     list_of_files = os.listdir(path)
     for file in list_of_files:
         old_file = os.path.join(path, file)
-        new_file = os.path.join(path, file.replace(old, new))
-        os.rename(old_file, new_file)
+        if os.path.isfile(old_file):
+            new_file = os.path.join(path, file.replace(old, new))
+            os.rename(old_file, new_file)
     return None
